create new module to house BIDS command utilities
